Made the Http client a subclass of Reuest
The Request alone only represents information about a request (to send or received).
An Http client is a request that can be sent. It creates a sender object to send the request
and receive the loaded content.
Usability is improved if Client is a subclass of Request (faster access to headers, params and url objects).